Reorganizes the admittedly small socket code. Adds a Readme.
