+++ src/server.ts
@@ -1,24 +1,14 @@
 import express from "express";
 import http from "http";
-import { Server } from "socket.io";
+import { initialize } from "./socket";
 
 const app = express();
 const server = new http.Server(app);
-const io = new Server(server, 
-  {
-    cors: {
-      origin: "*"
-    }
-  }
-);
-// const { PeerServer } = require(`peer`);
 
 app.use(express.static(`public`));
 
+
 app.get(`/`, (req, res) => {
-  // console.log((io.sockets.server.engine.clients))
-  // console.log((io.engine.clients))
-  // console.log(io.sockets.adapter.rooms.get("1234").size)
   res.status(200).json({ success: true })
 });
 
@@ -27,39 +17,6 @@ app.get(`/:room`, (req, res) => {
 });
 
 // Handles socket stuff
-io.on(`connection`, socket => {
-  // Listen event for joining a table
-  socket.on(`join-table`, (tableID: string, userID: string) => {
-    socket.join(tableID);
+initialize(server);
 
-    // console.log(socket.adapter.rooms.get(tableID));
-
-    // Informs other players of the newcomer. Informs newcomer of table size
-    const playersPresent = socket.adapter.rooms.get(tableID).size;
-    socket.to(tableID).broadcast.emit(`player-joined`, userID, playersPresent)
-    socket.emit(`i-joined`, playersPresent);
-
-    console.log(`User ${userID} has joined. ${playersPresent} players playing`)
-
-    socket.on(`disconnect`, () => {
-      const table = socket.adapter.rooms.get(tableID);
-      let playersPresent = 0;
-      if (table) { playersPresent = table.size; }
-      socket.to(tableID).broadcast.emit(`player-disconnected`, userID, playersPresent);
-      console.log(`User ${userID} has left. ${playersPresent} players remain`)
-    });
-  });
-
-  // Informs all players in the table that the current user is ready to play
-  socket.on(`ready`, (tableID: string, userID: string) => {
-    socket.to(tableID).broadcast.emit(`player-ready`, userID);
-  });
-
-  // Tells the players who the new host is. 
-  socket.on(`assume-host`, (tableID: string, userID: string) => {
-    socket.to(tableID).broadcast.emit(`new-host`, userID);
-  });
-});
-
-// peerServer = PeerServer({ port: 3002, path: `/peer` });
-server.listen(3001)
+server.listen(3001);
